refactor intergrate test
