+++ tests/integration_tests/crypto.rs
@@ -1,0 +1,50 @@
+use super::utils::hash;
+use bytes::Bytes;
+use overlord::Crypto;
+use std::error::Error;
+
+pub struct MockCrypto {
+    name: Bytes,
+}
+
+impl MockCrypto {
+    pub fn new(name: Bytes) -> Self {
+        MockCrypto { name }
+    }
+}
+
+impl Crypto for MockCrypto {
+    fn hash(&self, speech: Bytes) -> Bytes {
+        hash(&speech)
+    }
+
+    fn sign(&self, _hash: Bytes) -> Result<Bytes, Box<dyn Error + Send>> {
+        Ok(self.name.clone())
+    }
+
+    fn aggregate_signatures(
+        &self,
+        _signatures: Vec<Bytes>,
+        _speaker: Vec<Bytes>,
+    ) -> Result<Bytes, Box<dyn Error + Send>> {
+        Ok(Bytes::new())
+    }
+
+    fn verify_signature(
+        &self,
+        _signature: Bytes,
+        _hash: Bytes,
+        _voter: Bytes,
+    ) -> Result<(), Box<dyn Error + Send>> {
+        Ok(())
+    }
+
+    fn verify_aggregated_signature(
+        &self,
+        _aggregated_signature: Bytes,
+        _hash: Bytes,
+        _voters: Vec<Bytes>,
+    ) -> Result<(), Box<dyn Error + Send>> {
+        Ok(())
+    }
+}
